// apps/redux-toolkit/src/state/queries/models/spotify.models.ts
export type SpotifyDatePrecision = 'year' | 'month' | 'day';

export type SpotifyAccessToken = {
  access_token: string;
  token_type: string;
  scope: string;
  expires_in: number;
  refresh_token: string;
};

type SpotifyExternalUrls = {
  spotify: string;
};

type SpotifyExternalIds = {
  isrc: string;
  ean: string;
  upc: string;
};

type SpotifyFollowers = {
  href?: string;
  total: number;
};

type SpotifyImage = {
  url: string;
  height?: number;
  width?: number;
};

type SpotifyRestrictions = {
  reason: string;
};

export type SpotifyCopyRights = {
  text: string;
  type: string;
};

type SpotifyExplicitContent = {
  filter_enabled: boolean;
  filter_locked: boolean;
};

type SpotifyCursors = {
  after: string;
  before: string;
};

export type SpotifyArtist = {
  external_urls: SpotifyExternalUrls;
  followers: SpotifyFollowers;
  genres: string[];
  href: string;
  id: string;
  images: SpotifyImage[];
  name: string;
  popularity: number;
  type: 'artist';
  uri: string;
};
export type SimplifiedSpotifyArtist = Pick<
  SpotifyArtist,
  'external_urls' | 'href' | 'id' | 'name' | 'type' | 'uri'
>;

export type SpotifyAlbum = {
  album_type: 'album' | 'single' | 'compilation';
  total_tracks: number;
  available_markets: string[];
  external_urls: SpotifyExternalUrls;
  href: string;
  id: string;
  images: SpotifyImage[];
  name: string;
  release_date: string;
  release_date_precision: SpotifyDatePrecision;
  restrictions: SpotifyRestrictions;
  type: 'album';
  uri: string;
  artists: SimplifiedSpotifyArtist[];
  tracks: SpotifyAlbumTracks;
  copyrights: SpotifyCopyRights[];
  external_ids: SpotifyExternalIds;
  genres: string[];
  label: string;
  popularity: number;
};
export type SimplifiedSpotifyAlbum = Omit<
  SpotifyAlbum,
  'tracks' | 'copyrights' | 'external_ids' | 'genres' | 'label' | 'popularity'
> & { album_group: 'album' | 'single' | 'compilation' | 'appears_on' };

type SpotifyAlbumTracks = {
  href: string;
  limit: number;
  next: string | null;
  offset: number;
  previous: string | null;
  total: number;
  items: SimplifiedSpotifyTrack[];
};
// ...
export type SimplifiedSpotifyTrack = Omit<SpotifyTrack, 'album' | 'external_urls' | 'popularity'>;
// ...
export type DiscographyType = 'all' | 'album' | 'single' | 'compilation';